Move to etcd
Problem: This branch exists to test bundle support for CRDs provided by the
Prometheus operator. The operator we were deploying earlier was the
Prometheus operator, which does not scale well given that the CRDs
shouldn't be present prior to deploying the operator.

Solution: This commit updates the operator to etcd.